Set utf-8 output encoding
The save file doesn't have an explicit encoding.  Elsewhere, we assume
UTF-8 when parsing strings (the player and farmhand names).  So, let's
explicitly set utf-8 as the encoding when we save the file.  If we
didn't do this, lxml tries to save as ASCII; that would be a problem if
we actually encountered any non-ASCII names!