Support creating human-readable XML
lxml can create 'formatted' XML that is human-readable.  This adds that
as an option (which defaults to off).

At the same time, when the XML file is parsed, all human-readable
whitespace is removed.  This is needed because, if the input XML was
already human-readable (say, by running `xmllint --format` separately),
running this program without pretty-printing would leave at least some
of the pretty-printing in place, and could actually make it weird (that
is, it could mess up the tab spacing).

The side-effect of this is, if the user had applied pretty-printing
themselves, they have to remember to enable it when running this
program, else the pretty-printing will be removed.