Swap home locations
Each person has a `<homeLocation>` element containing the unique name of
their home location.  For the player it's `FarmHouse`, for the farmhand
it's the unique name of their cabin.  These two values need to be
swapped.

NOTE: This will definitely have follow-on effects, but I don't yet know
what those effects are.  But I bet one of the effects will hit when
someone is married.

Closes #1